requirement.txt
plotly
streamlit
pandas
mysql.connector
matplotlib

